Added unit tests for MapTile and GameMap
MapTile:
Changed the existing tests for the toString() method to instead be performed on the getSymbol() method which performs the same function as toString() used to perform. Then added tests to test the new toString() method which outputs more data.

GameMap:
Added tests for the new toString() and renderMap() methods. Both these methods may not be used much, but renderMap() is good for debugging, and toString() could see use in a save to disk function with some minor alterations (removing some of the text that helps with readability).