added dependency injection for the weapon damage calculation. We should also update our tests as described:
Weapon damage was generated with Math.random() internally.

part of the calculation looked like this: 1+Math.random() * 50.

Which opens up for a range of values for weapon damage.

What's the problem with that? Poor testing ability.

We may want to control the "Math.random()" part with dependency injection.

If we can instead choose a value, like 0.5, then we can test if the "1+..." and "*50" part are correct or not. If we have Math.random() instead, then we won't know for sure.

Furthermore the way we tested weapon damage, is by checking if the final value was within accepted range, for example 91-100. That sounds like a reasonable test case, right?

But what if the weapon can never have 91 in damage? Maybe it can only be 92 as minimum. Our tests may never show that because of Math.random().

If we think the lowest acceptable is 91, we need a test that checks if the damage 91 is possible or not. Plus another test to see that 90 fails, if that's what we expect.

I've already added a couple of examples in the equipment test class.